+++ cmd/motr/main.go
@@ -5,8 +5,8 @@ package main
 
 import (
 	"github.com/labstack/echo/v4"
-	"github.com/onsonr/sonr/cmd/motr/fetch"
 	"github.com/onsonr/sonr/internal/ctx"
+	"github.com/onsonr/sonr/internal/dwn"
 	"github.com/onsonr/sonr/pkg/workers/routes"
 )
 
@@ -14,5 +14,5 @@ func main() {
 	e := echo.New()
 	e.Use(ctx.SessionMiddleware)
 	routes.RegisterClientAPI(e)
-	fetch.Serve(e)
+	dwn.Serve(e)
 }
